fix all bug

+++ code/src/main/java/org/example/code/DTO/SanPhamChiTietDTO.java
@@ -2,6 +2,8 @@ package org.example.code.DTO;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SanPhamChiTietDTO {
     private Integer id;
@@ -10,7 +12,7 @@ public class SanPhamChiTietDTO {
     private Integer idPhienban;
     private String anhSP;
     private Integer soLuongTon;
-    private Double donGia;
+    private BigDecimal donGia;
     private String moTa;
     private String trangThai;
 }
