+++ code/src/main/java/org/example/code/model/HoaDonChiTiet.java
@@ -4,6 +4,8 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Entity
@@ -26,6 +28,6 @@ public class HoaDonChiTiet {
     private Integer soLuong;
 
     @Column(name = "donGia")
-    private Double donGia;
+    private BigDecimal donGia;
 
 }
